Fixed #1, #2 - off-model-route students can now register & build timetable online
The code to allow off the model route students to register for re-admission and to build timetables online is actually part of the initial commit. I just added the unit tests here.

The solution involves adding front-end controls to let a student indicate their intention of re-admission, and passing the relevant info to the back-end to register the student. Once this student's status looks like a regular student's, the system just treats them like any regular student, thus the student should be able to build their timetable.